reaftoracao do codigo de salvar compra no banco
