+++ src/Service/AuthenticationService.tsx
@@ -1,30 +1,5 @@
 import { save} from "./CookieService";
 
-interface Produto {
-    ped_id: number
-    cli_id:  number
-    pro_id: number
-    ped_status : string
-}
-
-export async function salvarCompra({ped_id,cli_id, pro_id, ped_status}: Produto) {
-
-    console.log('chegou na chamada')
-    const url = 'http://localhost:3000/actions/salvarCompra';
-
-    await fetch(url, {
-        method: 'POST',
-        headers: {
-            'Content-Type': 'application/json'
-        },
-        body: JSON.stringify({
-            ped_id,cli_id, pro_id, ped_status
-        })
-    }).catch((err) => err.message)
-
-}
-
-//localhost:3000/actions/salvarCompra
 export async function BuscarSessao(cli_token: string | null): Promise<string | null> {
     /*Para testes locais usar http://localhost:3000/cliente/autenticar/?cli_token=*/
